# specs/002-backoffice-catalog-sales/openapi.yaml
openapi: 3.0.3
info:
  title: Pede Aqui — Backoffice Catalog, Sales & Reports (Spec 002)
  version: 0.1.0
  description: >
    Contract fragment for Spec 002. Merge into the backend OpenAPI at /api/v1.
    All endpoints require Bearer JWT; tenant resolved from JWT tenant_id or X-Tenant-Id.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /sales/orders:
    get:
      tags: [sales]
      summary: Sales list (commercial lens over orders). VENDOR_ADMIN forced to own vendor.
      parameters:
        - { $ref: '#/components/parameters/TenantHeader' }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: status, in: query, schema: { type: string } }
        - { name: vendorId, in: query, schema: { type: string, format: uuid } }
        - { name: productId, in: query, schema: { type: string, format: uuid } }
        - { name: skuId, in: query, schema: { type: string, format: uuid } }
        - { name: paymentProvider, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/SalesPage' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
  /sales/orders/{orderId}:
    get:
      tags: [sales]
      summary: Sale detail (snapshots, discount, payments, refunds, commission).
      parameters: [{ $ref: '#/components/parameters/OrderId' }, { $ref: '#/components/parameters/TenantHeader' }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/SaleDetail' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /sales/orders/{orderId}/cancel:
    post:
      tags: [sales]
      summary: Cancel a sale via the order state machine (pre-dispatch states only).
      parameters: [{ $ref: '#/components/parameters/OrderId' }, { $ref: '#/components/parameters/TenantHeader' }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [reason], properties: { reason: { type: string, maxLength: 500 } } } } }
      responses:
        '200': { description: Cancelled }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: Not cancellable in current state }
  /sales/orders/{orderId}/refund:
    post:
      tags: [sales]
      summary: Full or partial refund via the finance/refund path. Idempotent.
      parameters:
        - { $ref: '#/components/parameters/OrderId' }
        - { $ref: '#/components/parameters/TenantHeader' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount: { type: number, format: double, description: "Omit for full refund; must be ≤ paid − alreadyRefunded" }
                reason: { type: string, maxLength: 500 }
      responses:
        '200': { description: Refund created }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { description: Invalid amount }
  /sales/orders/{orderId}/resend-notification:
    post:
      tags: [sales]
      summary: Re-send a customer notification. DELIVERY_CODE never returns/logs the OTP.
      parameters: [{ $ref: '#/components/parameters/OrderId' }, { $ref: '#/components/parameters/TenantHeader' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [type], properties: { type: { type: string, enum: [CONFIRMATION, STATUS, DELIVERY_CODE] } } }
      responses:
        '202': { description: Queued }
        '403': { $ref: '#/components/responses/Forbidden' }
  /sales/orders/{orderId}/status-override:
    post:
      tags: [sales]
      summary: Config-gated manual transition (allow-list only, ADMIN/OPS, reason required).
      parameters:
        - { $ref: '#/components/parameters/OrderId' }
        - { $ref: '#/components/parameters/TenantHeader' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [targetStatus, reason], properties: { targetStatus: { type: string }, reason: { type: string, maxLength: 500 } } }
      responses:
        '200': { description: Transitioned }
        '403': { description: Disabled or insufficient role }
        '409': { description: Transition not in allow-list }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    TenantHeader: { name: X-Tenant-Id, in: header, required: false, schema: { type: string, format: uuid } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: false, schema: { type: string } }
...
    OrderId: { name: orderId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
...
    Forbidden: { description: Insufficient role or cross-tenant/vendor access }
    NotFound: { description: Not found in tenant scope }
  schemas:
    Money: { type: number, format: double, description: MZN, 2 decimals }
...
    SalesRow:
      type: object
      properties:
        orderId: { type: string, format: uuid }
        reference: { type: string }
        createdAt: { type: string, format: date-time }
        vendorId: { type: string, format: uuid }
        vendorName: { type: string }
        itemCount: { type: integer }
        subtotal: { $ref: '#/components/schemas/Money' }
        fees: { $ref: '#/components/schemas/Money' }
        taxes: { $ref: '#/components/schemas/Money' }
        discountTotal: { $ref: '#/components/schemas/Money' }
        total: { $ref: '#/components/schemas/Money' }
        orderStatus: { type: string }
        paymentStatus: { type: string }
        paymentProvider: { type: string }
    SalesPage:
      type: object
      properties:
        content: { type: array, items: { $ref: '#/components/schemas/SalesRow' } }
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
    SaleDetail:
      allOf:
        - $ref: '#/components/schemas/SalesRow'
        - type: object
          properties:
            items:
              type: array
              items:
                type: object
                properties:
                  productNameSnapshot: { type: string }
                  unitPriceSnapshot: { $ref: '#/components/schemas/Money' }
                  quantity: { type: integer }
                  lineTotal: { $ref: '#/components/schemas/Money' }
            appliedPromotion: { allOf: [ { $ref: '#/components/schemas/Promotion' } ], nullable: true }
            refunds:
              type: array
              items: { type: object, properties: { id: { type: string, format: uuid }, amount: { $ref: '#/components/schemas/Money' }, status: { type: string } } }
            commission: { $ref: '#/components/schemas/Money' }
    PromotionUpsert:
      type: object
      required: [name, type, value, scope, startsAt, endsAt]
      properties:
        vendorId: { type: string, format: uuid, nullable: true }
        name: { type: string, maxLength: 140 }
        code: { type: string, maxLength: 40, nullable: true }
        type: { type: string, enum: [PERCENTAGE, FIXED_AMOUNT] }
        value: { type: number, format: double }
        scope: { type: string, enum: [ORDER, CATEGORY, PRODUCT] }
        targetCategoryId: { type: string, format: uuid, nullable: true }
        targetProductId: { type: string, format: uuid, nullable: true }
        minOrderTotal: { allOf: [ { $ref: '#/components/schemas/Money' } ], nullable: true }
        maxDiscountAmount: { allOf: [ { $ref: '#/components/schemas/Money' } ], nullable: true }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        usageLimit: { type: integer, nullable: true }
        perCustomerLimit: { type: integer, nullable: true }
    Promotion:
      allOf:
        - $ref: '#/components/schemas/PromotionUpsert'
        - type: object
          properties:
            id: { type: string, format: uuid }
            usedCount: { type: integer }
            status: { type: string, enum: [DRAFT, ACTIVE, PAUSED, EXPIRED] }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
...